#142 Split modules to monorepo packages
