--- .babelrc
@@ -1,3 +1,0 @@
-{
-  "extends": "./config/babel.config.js"
-}
