+++ .storybook/.babelrc
@@ -1,0 +1,4 @@
+{
+  "extends": "../babel.config.js",
+  "plugins": ["macros"]
+}
