+++ package.json
@@ -14,8 +14,9 @@
     "dev": "run-s dev:*",
     "test": "run-s test:*",
     "build": "run-s build:storybook",
-    "prepare": "run-s build",
-    "build:packages": "rollup --environment NODE_ENV:production -c",
+    "prepare": "run-s build:cjs build:esm",
+    "build:cjs": "yarn workspaces run rollup --environment NODE_ENV:production -f cjs -c",
+    "build:esm": "yarn workspaces run rollup --environment NODE_ENV:production -f esm -c",
     "build:storybook": "build-storybook -o dist",
     "dev:start": "start-storybook -p 9009 -s public",
     "test:eslint": "eslint --format unix ./src",
