+++ packages/content/rollup.config.js
@@ -1,3 +1,3 @@
-import config from '../../config/rollup.config';
+import config from '../../rollup.config';
 
 export default config;
