+++ packages/core/rollup.config.js
@@ -1,4 +1,6 @@
-import config from '../../config/rollup.config';
+import config from '../../rollup.config';
 
-export default config;
-
+export default {
+  ...config,
+  external: [...config.external, 'crypto'],
+};
