+++ packages/draft-editor/.babelrc
@@ -1,3 +1,3 @@
 {
-  "extends": "../../.babelrc"
+  "extends": "../../babel.config.js"
 }
