+++ packages/draft-editor/package.json
@@ -17,8 +17,5 @@
     "draft-js": "^0.11.0",
     "react": "^16.8.6",
     "styled-components": "^4.3.2"
-  },
-  "scripts": {
-    "build": "rollup --environment NODE_ENV:production -f esm -c"
   }
 }
