+++ packages/editor/package.json
@@ -20,8 +20,5 @@
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "styled-components": "^4.3.2"
-  },
-  "scripts": {
-    "build": "rollup --environment NODE_ENV:production -f esm -c"
   }
 }
