+++ packages/pie/package.json
@@ -13,8 +13,5 @@
   "peerDependencies": {
     "react": "^16.8.6",
     "styled-components": "^4.3.2"
-  },
-  "scripts": {
-    "build": "rollup --environment NODE_ENV:production -f esm -c"
   }
 }
